IUserAccountsProvider added and implemented.
